Allow upgrading pip; also yb_activate_virtualenv should fail if pip install fails (#11)
During virtualenv creation, if yb_virtualenv_upgrade_pip is set to "true", pip will be upgraded. If the virtualenv already exists, pip will never be upgraded.

We are not forcing pip upgrade onto environments where it is not needed. There might be advantages to using a predetermined version of pip in some environments.