Add better python venv 'correctness', allow specifying specific python version without a full path to the binary, also break code out into modules. (#15)
Currently (before this change) `yb_deactivate_virtualenv` has some deficiencies:
-  no check is made to see if the frozen file is newer than the requirements file 
-  if the venv dir already exists it is used as is.  No check is made to determine the python version it contains or to see if it is empty/incomplete (can happen is the task was aborted at just the right time)
-  no detection of other changes to env from when the requirements were frozen, like running on a different arch (arm vs x86)
-  no ability to specify specific python version which should be used aside from hard coding a full path to a python executable to use and then ensuring that every platform where this might run had the proper python version at that file path.
-  not easy to use in a standalone way (I would like to use this code everywhere, but having to pull in the entirety of yugabyte-bash-common makes that a little daunting)
-  switching between branches requires creating a new or updating existing venv, nothing is cached.
 
This PR aims to address these deficiencies.
-  Create a signature from the input to the pip command (requirements.txt, arch, python version) and add it as a comment to the frozen file.
-  Add a sentry file to the created venv that contains this signature.  A venv is determined 'up to date, ready to use' if the signature file is already present and the signature matches.  If the signature doesn't match a new venv is created.
- Add a check to verify the signature in the frozen file to see if it is valid (should catch arch and python changes).  If the signature doesn't match a new venv is created along with a new frozen file (useful for development when adding new modules)
- Add a check to verify that sentry file is the most recent file in the venv.  If any files are newer than the sentry file, we consider the venv invalid (been modified) and we create a new one.
- Add support for maintaining multiple venv outside the current directory (this is optional behavior) that allows for switching between branches of a repo and reusing an already existent venv.
- Add ability to set python version with arbitrary precision so long as the desired version is somewhere in $PATH
- Break out some basic functionality into modules to make it easier to use in other projects
- Add a stand alone create_venv.sh script for handy use from the shell.


Modules:
Python detection, logging, and venv creation have all been moved into their own modules.
Each module has a guard at the top to prevent it from being sourced more than once.

logger.sh: contains all the logging functions, including colors and timestamps.  This module has no external dependencies.  

detect_python.sh: detects what python to use based on YB_PYTHON_VERSION defaulting to 3.7.  It will error out if no python version greater than or equal to the version asked for is detected with the same major version.  For example, if we are asked for py3.7 (the default) but only py3.8 is found then py3.8 will be used.  But if py2.7 is requested and the underlying system has only py3, we error out.  The actual version of python used is saved in `yb_python_version_actual`.  All the variables used or set in this module get marked as readonly.  This modules has no external dependencies.

create_venv: This module can also be run as a script in which case it activates a venv in the current directory.  If sourced it provides `yb_activate_virtualenv` and `yb_deactivate_virtualenv`.  This module depends on logger.sh and detect_python.sh.
There are some changes in behavior introduced with how we handle venvs.  By default, a venv dir path is calculated based on a hash of the OS, python version, arch, and requirements.txt contents.  This allows us to reuse an existing venv if no changes have occurred in the environment and forces a new venv from scratch if anything has changed.  It also has controls for forcing early exit if the frozen requirements file is detected as being out of date or for recreating the frozen file if desired.  To preserve the old behavior of creating a venv in the top level next to the requirements.txt file export YB_USE_TOP_LEVEL_VENV=true

test/test.sh: remove top level import of `yugabyte-bash-common.sh`.  Tests run in subshells that import the module under test.